requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
apache-airflow
apache-airflow-core
apache-airflow-providers-amazon==9.8.0
apache-airflow-providers-common-compat==1.7.2
apache-airflow-providers-common-io==1.6.1
apache-airflow-providers-common-sql==1.27.3
apache-airflow-providers-http==5.3.2
apache-airflow-providers-smtp==2.1.1
apache-airflow-providers-snowflake==6.3.1
apache-airflow-providers-standard==1.4.1
apache-airflow-task-sdk==1.0.3
argcomplete==3.6.2
asgiref==3.9.1
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.3.0
backports-datetime-fromisoformat==2.0.3
backports.strenum==1.3.1
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.39.14
botocore==1.39.14
cadwyn==5.4.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
colorlog==6.9.0
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.5
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
filelock==3.18.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.7.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonpath-ng==1.7.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
libcst==1.8.2
linkify-it-py==2.0.3
lockfile==0.12.2
lxml>=5.3.2,<5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.6.3
mypy_extensions==1.1.0
numpy==1.26.4
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
packaging==25.0
pandas==2.1.4
pandera==0.25.0
pathspec==0.12.1
pendulum==3.1.0
platformdirs==4.3.8
pluggy==1.6.0
ply==3.11
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pyarrow==18.1.0
PyAthena==3.15.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytest
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redshift-connector==2.1.8
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
retryhttp==1.3.3
rich==14.1.0
rich-argparse==1.7.1
rich-toolkit==0.14.8
rignore==0.6.4
rpds-py==0.26.0
s3transfer==0.13.1
sagemaker_studio==1.0.20
scramp==1.4.6
sentry-sdk==2.33.2
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.16.0
snowflake-snowpark-python==1.35.0
snowflake-sqlalchemy==1.7.6
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.47.2
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
tomli==2.2.1
tomlkit==0.13.3
typeguard==4.4.4
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
watchtower==3.4.0
websockets==15.0.1
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
